+++ contrib/libpodpy/__init__.py
@@ -1,0 +1,4 @@
+
+
+#__version__ = version
+__title__ = 'libpod'
