+++ contrib/libpodpy/containers.py
@@ -1,0 +1,8 @@
+
+class Containers(object):
+
+    def __init__(self, client):
+        self.client = client
+
+    def List(self):
+        pass
