+++ pkg/varlinkapi/util.go
@@ -1,0 +1,10 @@
+package varlinkapi
+
+import (
+	"context"
+)
+
+// getContext returns a non-nil, empty context
+func getContext() context.Context {
+	return context.TODO()
+}
